Refactor to enable the old style nexus plugin as well
