+++ src/main/groovy/se/alipsa/gradle/plugin/release/PublishingType.groovy
@@ -1,0 +1,5 @@
+package se.alipsa.gradle.plugin.release
+
+enum PublishingType {
+  NEXUS, CENTRAL
+}
